Distribute tests to match package structure
