+++ tests/test_mnist.py
@@ -1,8 +1,5 @@
-import sys
-print(sys.path)
-
-import pytest
 from applications.learn_mnist import train_and_evaluate_mnist
+
 
 def test_mnist_accuracy():
     test_accuracy = train_and_evaluate_mnist()
